Add empty tmp directory.
The web application provisioner assumes that there exists an `tmp` file
at the root of the repository to stage files. Rather than check if this
exists and create it if not, the `empty` file ensures that the correct
directory structure exists in version control for others who want to use
the repository.