Add startup delay to clickhouse instances.
The clickhouse instances have a large amount of data to load into memory
for the genetics DB. The Terraform healthcheck is just looking if the
Clickhouse instance is running, so the infrastructure can think that the
DB is ready for connections before it is. When this happens the loading
of the data can be corrupted. This shows in query errors as columns
being missing from some tables.